requirements.txt
praw
pandas
numpy
scikit-learn
nltk
yfinance
python-dotenv
flask
werkzeug
matplotlib
seaborn
textblob
sentence-transformers
xgboost